+++ boluo/src/listener.rs
@@ -4,7 +4,7 @@ use std::net::SocketAddr;
 
 /// 连接信息。
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub struct ConnectionInfo {
+pub struct ConnectInfo {
     /// 连接的本地地址。
     pub local: SocketAddr,
     /// 连接的远程地址。
@@ -31,7 +31,7 @@ pub trait Listener {
 #[cfg(feature = "tokio")]
 impl Listener for tokio::net::TcpListener {
     type IO = tokio::net::TcpStream;
-    type Addr = ConnectionInfo;
+    type Addr = ConnectInfo;
     type Error = std::io::Error;
 
     async fn accept(&mut self) -> std::io::Result<(Self::IO, Self::Addr)> {
@@ -39,7 +39,7 @@ impl Listener for tokio::net::TcpListener {
             .await
             .and_then(|(conn, remote)| {
                 self.local_addr()
-                    .map(|local| (conn, ConnectionInfo { local, remote }))
+                    .map(|local| (conn, ConnectInfo { local, remote }))
             })
     }
 }